AI/github_models/requirements.txt
openai
openai-agents
python-dotenv